+++ sdk/hdinsight/azure-mgmt-hdinsight/azure/mgmt/hdinsight/models/_models_py3.py
@@ -128,19 +128,23 @@ class ApplicationGetEndpoint(Model):
     :type destination_port: int
     :param public_port: The public port to connect to.
     :type public_port: int
+    :param private_ip_address: The private ip address of the endpoint.
+    :type private_ip_address: str
     """
 
     _attribute_map = {
         'location': {'key': 'location', 'type': 'str'},
         'destination_port': {'key': 'destinationPort', 'type': 'int'},
         'public_port': {'key': 'publicPort', 'type': 'int'},
+        'private_ip_address': {'key': 'privateIPAddress', 'type': 'str'},
     }
 
-    def __init__(self, *, location: str=None, destination_port: int=None, public_port: int=None, **kwargs) -> None:
+    def __init__(self, *, location: str=None, destination_port: int=None, public_port: int=None, private_ip_address: str=None, **kwargs) -> None:
         super(ApplicationGetEndpoint, self).__init__(**kwargs)
         self.location = location
         self.destination_port = destination_port
         self.public_port = public_port
+        self.private_ip_address = private_ip_address
 
 
 class ApplicationGetHttpsEndpoint(Model):
@@ -426,29 +430,46 @@ class BillingResponseListResult(Model):
     """The response for the operation to get regional billingSpecs for a
     subscription.
 
+    Variables are only populated by the server, and will be ignored when
+    sending a request.
+
     :param vm_sizes: The virtual machine sizes to include or exclude.
     :type vm_sizes: list[str]
+    :param vm_sizes_with_encryption_at_host: The vm sizes which enable
+     encryption at host.
+    :type vm_sizes_with_encryption_at_host: list[str]
     :param vm_size_filters: The virtual machine filtering mode. Effectively
      this can enabling or disabling the virtual machine sizes in a particular
      set.
     :type vm_size_filters:
      list[~azure.mgmt.hdinsight.models.VmSizeCompatibilityFilterV2]
+    :ivar vm_size_properties: The vm size properties.
+    :vartype vm_size_properties:
+     list[~azure.mgmt.hdinsight.models.VmSizeProperty]
     :param billing_resources: The billing and managed disk billing resources
      for a region.
     :type billing_resources:
      list[~azure.mgmt.hdinsight.models.BillingResources]
     """
 
+    _validation = {
+        'vm_size_properties': {'readonly': True},
+    }
+
     _attribute_map = {
         'vm_sizes': {'key': 'vmSizes', 'type': '[str]'},
+        'vm_sizes_with_encryption_at_host': {'key': 'vmSizesWithEncryptionAtHost', 'type': '[str]'},
         'vm_size_filters': {'key': 'vmSizeFilters', 'type': '[VmSizeCompatibilityFilterV2]'},
+        'vm_size_properties': {'key': 'vmSizeProperties', 'type': '[VmSizeProperty]'},
         'billing_resources': {'key': 'billingResources', 'type': '[BillingResources]'},
     }
 
-    def __init__(self, *, vm_sizes=None, vm_size_filters=None, billing_resources=None, **kwargs) -> None:
+    def __init__(self, *, vm_sizes=None, vm_sizes_with_encryption_at_host=None, vm_size_filters=None, billing_resources=None, **kwargs) -> None:
         super(BillingResponseListResult, self).__init__(**kwargs)
         self.vm_sizes = vm_sizes
+        self.vm_sizes_with_encryption_at_host = vm_sizes_with_encryption_at_host
         self.vm_size_filters = vm_size_filters
+        self.vm_size_properties = None
         self.billing_resources = billing_resources
 
 
@@ -781,6 +802,8 @@ class ClusterGetProperties(Model):
 
     :param cluster_version: The version of the cluster.
     :type cluster_version: str
+    :param cluster_hdp_version: The hdp version of the cluster.
+    :type cluster_hdp_version: str
     :param os_type: The type of operating system. Possible values include:
      'Windows', 'Linux'
     :type os_type: str or ~azure.mgmt.hdinsight.models.OSType
@@ -821,8 +844,13 @@ class ClusterGetProperties(Model):
      properties.
     :type encryption_in_transit_properties:
      ~azure.mgmt.hdinsight.models.EncryptionInTransitProperties
+    :param storage_profile: The storage profile.
+    :type storage_profile: ~azure.mgmt.hdinsight.models.StorageProfile
     :param min_supported_tls_version: The minimal supported tls version.
     :type min_supported_tls_version: str
+    :param excluded_services_config: The excluded services config.
+    :type excluded_services_config:
+     ~azure.mgmt.hdinsight.models.ExcludedServicesConfig
     :param network_properties: The network properties.
     :type network_properties: ~azure.mgmt.hdinsight.models.NetworkProperties
     :param compute_isolation_properties: The compute isolation properties.
@@ -836,6 +864,7 @@ class ClusterGetProperties(Model):
 
     _attribute_map = {
         'cluster_version': {'key': 'clusterVersion', 'type': 'str'},
+        'cluster_hdp_version': {'key': 'clusterHdpVersion', 'type': 'str'},
         'os_type': {'key': 'osType', 'type': 'OSType'},
         'tier': {'key': 'tier', 'type': 'Tier'},
         'cluster_id': {'key': 'clusterId', 'type': 'str'},
@@ -851,14 +880,17 @@ class ClusterGetProperties(Model):
         'connectivity_endpoints': {'key': 'connectivityEndpoints', 'type': '[ConnectivityEndpoint]'},
         'disk_encryption_properties': {'key': 'diskEncryptionProperties', 'type': 'DiskEncryptionProperties'},
         'encryption_in_transit_properties': {'key': 'encryptionInTransitProperties', 'type': 'EncryptionInTransitProperties'},
+        'storage_profile': {'key': 'storageProfile', 'type': 'StorageProfile'},
         'min_supported_tls_version': {'key': 'minSupportedTlsVersion', 'type': 'str'},
+        'excluded_services_config': {'key': 'excludedServicesConfig', 'type': 'ExcludedServicesConfig'},
         'network_properties': {'key': 'networkProperties', 'type': 'NetworkProperties'},
         'compute_isolation_properties': {'key': 'computeIsolationProperties', 'type': 'ComputeIsolationProperties'},
     }
 
-    def __init__(self, *, cluster_definition, cluster_version: str=None, os_type=None, tier=None, cluster_id: str=None, kafka_rest_properties=None, security_profile=None, compute_profile=None, provisioning_state=None, created_date: str=None, cluster_state: str=None, quota_info=None, errors=None, connectivity_endpoints=None, disk_encryption_properties=None, encryption_in_transit_properties=None, min_supported_tls_version: str=None, network_properties=None, compute_isolation_properties=None, **kwargs) -> None:
+    def __init__(self, *, cluster_definition, cluster_version: str=None, cluster_hdp_version: str=None, os_type=None, tier=None, cluster_id: str=None, kafka_rest_properties=None, security_profile=None, compute_profile=None, provisioning_state=None, created_date: str=None, cluster_state: str=None, quota_info=None, errors=None, connectivity_endpoints=None, disk_encryption_properties=None, encryption_in_transit_properties=None, storage_profile=None, min_supported_tls_version: str=None, excluded_services_config=None, network_properties=None, compute_isolation_properties=None, **kwargs) -> None:
         super(ClusterGetProperties, self).__init__(**kwargs)
         self.cluster_version = cluster_version
+        self.cluster_hdp_version = cluster_hdp_version
         self.os_type = os_type
         self.tier = tier
         self.cluster_id = cluster_id
@@ -874,7 +906,9 @@ class ClusterGetProperties(Model):
         self.connectivity_endpoints = connectivity_endpoints
         self.disk_encryption_properties = disk_encryption_properties
         self.encryption_in_transit_properties = encryption_in_transit_properties
+        self.storage_profile = storage_profile
         self.min_supported_tls_version = min_supported_tls_version
+        self.excluded_services_config = excluded_services_config
         self.network_properties = network_properties
         self.compute_isolation_properties = compute_isolation_properties
 
@@ -934,6 +968,8 @@ class ClusterIdentityUserAssignedIdentitiesValue(Model):
     :vartype principal_id: str
     :ivar client_id: The client id of user assigned identity.
     :vartype client_id: str
+    :param tenant_id: The tenant id of user assigned identity.
+    :type tenant_id: str
     """
 
     _validation = {
@@ -944,12 +980,14 @@ class ClusterIdentityUserAssignedIdentitiesValue(Model):
     _attribute_map = {
         'principal_id': {'key': 'principalId', 'type': 'str'},
         'client_id': {'key': 'clientId', 'type': 'str'},
+        'tenant_id': {'key': 'tenantId', 'type': 'str'},
     }
 
-    def __init__(self, **kwargs) -> None:
+    def __init__(self, *, tenant_id: str=None, **kwargs) -> None:
         super(ClusterIdentityUserAssignedIdentitiesValue, self).__init__(**kwargs)
         self.principal_id = None
         self.client_id = None
+        self.tenant_id = tenant_id
 
 
 class ClusterListPersistedScriptActionsResult(Model):
@@ -1130,6 +1168,8 @@ class ConnectivityEndpoint(Model):
     :type location: str
     :param port: The port to connect to.
     :type port: int
+    :param private_ip_address: The private ip address of the endpoint.
+    :type private_ip_address: str
     """
 
     _attribute_map = {
@@ -1137,14 +1177,16 @@ class ConnectivityEndpoint(Model):
         'protocol': {'key': 'protocol', 'type': 'str'},
         'location': {'key': 'location', 'type': 'str'},
         'port': {'key': 'port', 'type': 'int'},
+        'private_ip_address': {'key': 'privateIPAddress', 'type': 'str'},
     }
 
-    def __init__(self, *, name: str=None, protocol: str=None, location: str=None, port: int=None, **kwargs) -> None:
+    def __init__(self, *, name: str=None, protocol: str=None, location: str=None, port: int=None, private_ip_address: str=None, **kwargs) -> None:
         super(ConnectivityEndpoint, self).__init__(**kwargs)
         self.name = name
         self.protocol = protocol
         self.location = location
         self.port = port
+        self.private_ip_address = private_ip_address
 
 
 class DataDisksGroups(Model):
@@ -1178,6 +1220,35 @@ class DataDisksGroups(Model):
         self.disks_per_node = disks_per_node
         self.storage_account_type = None
         self.disk_size_gb = None
+
+
+class Dimension(Model):
+    """The definition of Dimension.
+
+    :param name: The name of the dimension.
+    :type name: str
+    :param display_name: The display name of the dimension.
+    :type display_name: str
+    :param internal_name: The display name of the dimension.
+    :type internal_name: str
+    :param to_be_exported_for_shoebox: The flag indicates whether the metric
+     will be exported for shoebox or not.
+    :type to_be_exported_for_shoebox: bool
+    """
+
+    _attribute_map = {
+        'name': {'key': 'name', 'type': 'str'},
+        'display_name': {'key': 'displayName', 'type': 'str'},
+        'internal_name': {'key': 'internalName', 'type': 'str'},
+        'to_be_exported_for_shoebox': {'key': 'toBeExportedForShoebox', 'type': 'bool'},
+    }
+
+    def __init__(self, *, name: str=None, display_name: str=None, internal_name: str=None, to_be_exported_for_shoebox: bool=None, **kwargs) -> None:
+        super(Dimension, self).__init__(**kwargs)
+        self.name = name
+        self.display_name = display_name
+        self.internal_name = internal_name
+        self.to_be_exported_for_shoebox = to_be_exported_for_shoebox
 
 
 class DiskBillingMeters(Model):
@@ -1314,6 +1385,26 @@ class Errors(Model):
         super(Errors, self).__init__(**kwargs)
         self.code = code
         self.message = message
+
+
+class ExcludedServicesConfig(Model):
+    """The configuration that services will be excluded when creating cluster.
+
+    :param excluded_services_config_id: The config id of excluded services.
+    :type excluded_services_config_id: str
+    :param excluded_services_list: The list of excluded services.
+    :type excluded_services_list: str
+    """
+
+    _attribute_map = {
+        'excluded_services_config_id': {'key': 'excludedServicesConfigId', 'type': 'str'},
+        'excluded_services_list': {'key': 'excludedServicesList', 'type': 'str'},
+    }
+
+    def __init__(self, *, excluded_services_config_id: str=None, excluded_services_list: str=None, **kwargs) -> None:
+        super(ExcludedServicesConfig, self).__init__(**kwargs)
+        self.excluded_services_config_id = excluded_services_config_id
+        self.excluded_services_list = excluded_services_list
 
 
 class ExecuteScriptActionParameters(Model):
@@ -1501,6 +1592,94 @@ class LocalizedName(Model):
         self.localized_value = localized_value
 
 
+class MetricSpecifications(Model):
+    """The details of metric specifications.
+
+    :param name: The name of the metric specification.
+    :type name: str
+    :param display_name: The display name of the metric specification.
+    :type display_name: str
+    :param display_description: The display description of the metric
+     specification.
+    :type display_description: str
+    :param unit: The unit of the metric specification.
+    :type unit: str
+    :param aggregation_type: The aggregation type of the metric specification.
+    :type aggregation_type: str
+    :param supported_aggregation_types: The supported aggregation types of the
+     metric specification.
+    :type supported_aggregation_types: list[str]
+    :param supported_time_grain_types: The supported time grain types of the
+     metric specification.
+    :type supported_time_grain_types: list[str]
+    :param enable_regional_mdm_account: The flag indicates whether enable
+     regional mdm account or not.
+    :type enable_regional_mdm_account: bool
+    :param source_mdm_account: The source mdm account.
+    :type source_mdm_account: str
+    :param source_mdm_namespace: The source mdm namespace.
+    :type source_mdm_namespace: str
+    :param metric_filter_pattern: The metric filter pattern.
+    :type metric_filter_pattern: str
+    :param fill_gap_with_zero: The flag indicates whether filling gap with
+     zero.
+    :type fill_gap_with_zero: bool
+    :param category: The category of the metric.
+    :type category: str
+    :param resource_id_dimension_name_override: The override name of resource
+     id dimension name.
+    :type resource_id_dimension_name_override: str
+    :param is_internal: The flag indicates whether the metric is internal or
+     not.
+    :type is_internal: bool
+    :param delegate_metric_name_override: The override name of delegate
+     metric.
+    :type delegate_metric_name_override: str
+    :param dimensions: The dimensions of the metric specification.
+    :type dimensions: list[~azure.mgmt.hdinsight.models.Dimension]
+    """
+
+    _attribute_map = {
+        'name': {'key': 'name', 'type': 'str'},
+        'display_name': {'key': 'displayName', 'type': 'str'},
+        'display_description': {'key': 'displayDescription', 'type': 'str'},
+        'unit': {'key': 'unit', 'type': 'str'},
+        'aggregation_type': {'key': 'aggregationType', 'type': 'str'},
+        'supported_aggregation_types': {'key': 'supportedAggregationTypes', 'type': '[str]'},
+        'supported_time_grain_types': {'key': 'supportedTimeGrainTypes', 'type': '[str]'},
+        'enable_regional_mdm_account': {'key': 'enableRegionalMdmAccount', 'type': 'bool'},
+        'source_mdm_account': {'key': 'sourceMdmAccount', 'type': 'str'},
+        'source_mdm_namespace': {'key': 'sourceMdmNamespace', 'type': 'str'},
+        'metric_filter_pattern': {'key': 'metricFilterPattern', 'type': 'str'},
+        'fill_gap_with_zero': {'key': 'fillGapWithZero', 'type': 'bool'},
+        'category': {'key': 'category', 'type': 'str'},
+        'resource_id_dimension_name_override': {'key': 'resourceIdDimensionNameOverride', 'type': 'str'},
+        'is_internal': {'key': 'isInternal', 'type': 'bool'},
+        'delegate_metric_name_override': {'key': 'delegateMetricNameOverride', 'type': 'str'},
+        'dimensions': {'key': 'dimensions', 'type': '[Dimension]'},
+    }
+
+    def __init__(self, *, name: str=None, display_name: str=None, display_description: str=None, unit: str=None, aggregation_type: str=None, supported_aggregation_types=None, supported_time_grain_types=None, enable_regional_mdm_account: bool=None, source_mdm_account: str=None, source_mdm_namespace: str=None, metric_filter_pattern: str=None, fill_gap_with_zero: bool=None, category: str=None, resource_id_dimension_name_override: str=None, is_internal: bool=None, delegate_metric_name_override: str=None, dimensions=None, **kwargs) -> None:
+        super(MetricSpecifications, self).__init__(**kwargs)
+        self.name = name
+        self.display_name = display_name
+        self.display_description = display_description
+        self.unit = unit
+        self.aggregation_type = aggregation_type
+        self.supported_aggregation_types = supported_aggregation_types
+        self.supported_time_grain_types = supported_time_grain_types
+        self.enable_regional_mdm_account = enable_regional_mdm_account
+        self.source_mdm_account = source_mdm_account
+        self.source_mdm_namespace = source_mdm_namespace
+        self.metric_filter_pattern = metric_filter_pattern
+        self.fill_gap_with_zero = fill_gap_with_zero
+        self.category = category
+        self.resource_id_dimension_name_override = resource_id_dimension_name_override
+        self.is_internal = is_internal
+        self.delegate_metric_name_override = delegate_metric_name_override
+        self.dimensions = dimensions
+
+
 class NetworkProperties(Model):
     """The network properties.
 
@@ -1531,17 +1710,21 @@ class Operation(Model):
     :type name: str
     :param display: The object that represents the operation.
     :type display: ~azure.mgmt.hdinsight.models.OperationDisplay
+    :param properties: The operation properties.
+    :type properties: ~azure.mgmt.hdinsight.models.OperationProperties
     """
 
     _attribute_map = {
         'name': {'key': 'name', 'type': 'str'},
         'display': {'key': 'display', 'type': 'OperationDisplay'},
+        'properties': {'key': 'properties', 'type': 'OperationProperties'},
     }
 
-    def __init__(self, *, name: str=None, display=None, **kwargs) -> None:
+    def __init__(self, *, name: str=None, display=None, properties=None, **kwargs) -> None:
         super(Operation, self).__init__(**kwargs)
         self.name = name
         self.display = display
+        self.properties = properties
 
 
 class OperationDisplay(Model):
@@ -1554,19 +1737,40 @@ class OperationDisplay(Model):
     :type resource: str
     :param operation: The operation type: read, write, delete, etc.
     :type operation: str
+    :param description: Localized friendly description for the operation
+    :type description: str
     """
 
     _attribute_map = {
         'provider': {'key': 'provider', 'type': 'str'},
         'resource': {'key': 'resource', 'type': 'str'},
         'operation': {'key': 'operation', 'type': 'str'},
+        'description': {'key': 'description', 'type': 'str'},
     }
 
-    def __init__(self, *, provider: str=None, resource: str=None, operation: str=None, **kwargs) -> None:
+    def __init__(self, *, provider: str=None, resource: str=None, operation: str=None, description: str=None, **kwargs) -> None:
         super(OperationDisplay, self).__init__(**kwargs)
         self.provider = provider
         self.resource = resource
         self.operation = operation
+        self.description = description
+
+
+class OperationProperties(Model):
+    """The details of operation.
+
+    :param service_specification: The specification of the service.
+    :type service_specification:
+     ~azure.mgmt.hdinsight.models.ServiceSpecification
+    """
+
+    _attribute_map = {
+        'service_specification': {'key': 'serviceSpecification', 'type': 'ServiceSpecification'},
+    }
+
+    def __init__(self, *, service_specification=None, **kwargs) -> None:
+        super(OperationProperties, self).__init__(**kwargs)
+        self.service_specification = service_specification
 
 
 class OperationResource(Model):
@@ -1712,6 +1916,8 @@ class Role(Model):
      list[~azure.mgmt.hdinsight.models.DataDisksGroups]
     :param script_actions: The list of script actions on the role.
     :type script_actions: list[~azure.mgmt.hdinsight.models.ScriptAction]
+    :param encrypt_data_disks: Indicates whether encrypt the data disks.
+    :type encrypt_data_disks: bool
     """
 
     _attribute_map = {
@@ -1724,9 +1930,10 @@ class Role(Model):
         'virtual_network_profile': {'key': 'virtualNetworkProfile', 'type': 'VirtualNetworkProfile'},
         'data_disks_groups': {'key': 'dataDisksGroups', 'type': '[DataDisksGroups]'},
         'script_actions': {'key': 'scriptActions', 'type': '[ScriptAction]'},
+        'encrypt_data_disks': {'key': 'encryptDataDisks', 'type': 'bool'},
     }
 
-    def __init__(self, *, name: str=None, min_instance_count: int=None, target_instance_count: int=None, autoscale_configuration=None, hardware_profile=None, os_profile=None, virtual_network_profile=None, data_disks_groups=None, script_actions=None, **kwargs) -> None:
+    def __init__(self, *, name: str=None, min_instance_count: int=None, target_instance_count: int=None, autoscale_configuration=None, hardware_profile=None, os_profile=None, virtual_network_profile=None, data_disks_groups=None, script_actions=None, encrypt_data_disks: bool=None, **kwargs) -> None:
         super(Role, self).__init__(**kwargs)
         self.name = name
         self.min_instance_count = min_instance_count
@@ -1737,6 +1944,7 @@ class Role(Model):
         self.virtual_network_profile = virtual_network_profile
         self.data_disks_groups = data_disks_groups
         self.script_actions = script_actions
+        self.encrypt_data_disks = encrypt_data_disks
 
 
 class RuntimeScriptAction(Model):
@@ -2005,6 +2213,23 @@ class SecurityProfile(Model):
         self.cluster_users_group_dns = cluster_users_group_dns
         self.aadds_resource_id = aadds_resource_id
         self.msi_resource_id = msi_resource_id
+
+
+class ServiceSpecification(Model):
+    """The specification of the service.
+
+    :param metric_specifications: The metric specifications.
+    :type metric_specifications:
+     list[~azure.mgmt.hdinsight.models.MetricSpecifications]
+    """
+
+    _attribute_map = {
+        'metric_specifications': {'key': 'metricSpecifications', 'type': '[MetricSpecifications]'},
+    }
+
+    def __init__(self, *, metric_specifications=None, **kwargs) -> None:
+        super(ServiceSpecification, self).__init__(**kwargs)
+        self.metric_specifications = metric_specifications
 
 
 class SshProfile(Model):
@@ -2280,7 +2505,7 @@ class VmSizeCompatibilityFilterV2(Model):
 
     :param filter_mode: The filtering mode. Effectively this can enabling or
      disabling the VM sizes in a particular set. Possible values include:
-     'Exclude', 'Include'
+     'Exclude', 'Include', 'Recommend', 'Default'
     :type filter_mode: str or ~azure.mgmt.hdinsight.models.FilterMode
     :param regions: The list of regions under the effect of the filter.
     :type regions: list[str]
@@ -2317,6 +2542,62 @@ class VmSizeCompatibilityFilterV2(Model):
         self.cluster_versions = cluster_versions
         self.os_type = os_type
         self.vm_sizes = vm_sizes
+
+
+class VmSizeProperty(Model):
+    """The vm size property.
+
+    :param name: The vm size name.
+    :type name: str
+    :param cores: The number of cores that the vm size has.
+    :type cores: int
+    :param data_disk_storage_tier: The data disk storage tier of the vm size.
+    :type data_disk_storage_tier: str
+    :param label: The label of the vm size.
+    :type label: str
+    :param max_data_disk_count: The max data disk count of the vm size.
+    :type max_data_disk_count: long
+    :param memory_in_mb: The memory whose unit is MB of the vm size.
+    :type memory_in_mb: long
+    :param supported_by_virtual_machines: This indicates this vm size is
+     supported by virtual machines or not
+    :type supported_by_virtual_machines: bool
+    :param supported_by_web_worker_roles: The indicates this vm size is
+     supported by web worker roles or not
+    :type supported_by_web_worker_roles: bool
+    :param virtual_machine_resource_disk_size_in_mb: The virtual machine
+     resource disk size whose unit is MB of the vm size.
+    :type virtual_machine_resource_disk_size_in_mb: long
+    :param web_worker_resource_disk_size_in_mb: The web worker resource disk
+     size whose unit is MB of the vm size.
+    :type web_worker_resource_disk_size_in_mb: long
+    """
+
+    _attribute_map = {
+        'name': {'key': 'name', 'type': 'str'},
+        'cores': {'key': 'cores', 'type': 'int'},
+        'data_disk_storage_tier': {'key': 'dataDiskStorageTier', 'type': 'str'},
+        'label': {'key': 'label', 'type': 'str'},
+        'max_data_disk_count': {'key': 'maxDataDiskCount', 'type': 'long'},
+        'memory_in_mb': {'key': 'memoryInMb', 'type': 'long'},
+        'supported_by_virtual_machines': {'key': 'supportedByVirtualMachines', 'type': 'bool'},
+        'supported_by_web_worker_roles': {'key': 'supportedByWebWorkerRoles', 'type': 'bool'},
+        'virtual_machine_resource_disk_size_in_mb': {'key': 'virtualMachineResourceDiskSizeInMb', 'type': 'long'},
+        'web_worker_resource_disk_size_in_mb': {'key': 'webWorkerResourceDiskSizeInMb', 'type': 'long'},
+    }
+
+    def __init__(self, *, name: str=None, cores: int=None, data_disk_storage_tier: str=None, label: str=None, max_data_disk_count: int=None, memory_in_mb: int=None, supported_by_virtual_machines: bool=None, supported_by_web_worker_roles: bool=None, virtual_machine_resource_disk_size_in_mb: int=None, web_worker_resource_disk_size_in_mb: int=None, **kwargs) -> None:
+        super(VmSizeProperty, self).__init__(**kwargs)
+        self.name = name
+        self.cores = cores
+        self.data_disk_storage_tier = data_disk_storage_tier
+        self.label = label
+        self.max_data_disk_count = max_data_disk_count
+        self.memory_in_mb = memory_in_mb
+        self.supported_by_virtual_machines = supported_by_virtual_machines
+        self.supported_by_web_worker_roles = supported_by_web_worker_roles
+        self.virtual_machine_resource_disk_size_in_mb = virtual_machine_resource_disk_size_in_mb
+        self.web_worker_resource_disk_size_in_mb = web_worker_resource_disk_size_in_mb
 
 
 class VmSizesCapability(Model):